# Speedup Firefox Nightly startup on MacOS with devtools protocol