… packages/devtools/tests/__snapshots__/launcher.test.js.snap