## Proposed changes

Speedup firefox/edge startup on MacOS
fixes #5916

results on my machine:
- before 5.5s
- after 1.5s

If the browser installed to the default location no waiting is required!

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation update

## Checklist

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] I have added proper type definitions for new commands (if appropriate)

## Further comments

Decided to use JSON format intentionally, ti takes ~300ms less than text and ~100ms less than XML output types.
Sorry to say there are no other reliable ways to get an app installation path on MacOS.

### Reviewers: @webdriverio/project-committers