@joventuraz 2020-10-02
This also fixes an issue I encountered after installing firefox through brew cask, where it was not added to Launch Services yet, so the lsregister command returned nothing.